Bouton de suppresion pour capitaine equipe

+++ resources/views/equipe/profil.blade.php
@@ -17,7 +17,11 @@
                     <th scope="col">Grade</th>
                     <th scope="col"></th>
                     <th scope="col">Pseudo</th>
-                    <th scope="col"></th>
+                    @if (Auth::check())
+                        @if (Auth::user()->id == $equipe->id_capitaine)
+                            <th scope="col"></th>
+                        @endif
+                    @endif
                 </tr>
             </thead>
             <tbody>
@@ -34,7 +38,11 @@
                             <th scope="row">
                                 <a href="/joueurs/{{$joueur->pseudo}}" class="text-white">{{ $joueur->pseudo }}</a>  
                             </th>
-                            <th></th>
+                            @if (Auth::check())
+                                @if (Auth::user()->id == $equipe->id_capitaine)
+                                    <th></th>
+                                @endif
+                            @endif
                         </tr>
                     @else
                         <tr>
@@ -48,9 +56,14 @@
                             <th scope="row">
                                 <a href="/joueurs/{{$joueur->pseudo}}" class="text-white">{{ $joueur->pseudo }}</a>  
                             </th>
-                            <th>
-                                <a href="/equipes/{{$equipe->id}}/delete/joueur/{{$joueur->id_user}}" class="btn btn-danger">Supprimer</a>
-                            </th>
+                            
+                            @if (Auth::check())
+                                @if (Auth::user()->id == $equipe->id_capitaine)
+                                    <th>
+                                        <a href="/equipes/{{$equipe->id}}/delete/joueur/{{$joueur->id_user}}" class="btn btn-danger">Supprimer</a>
+                                    </th>
+                                @endif
+                            @endif
                         </tr>
                     @endif
                 @endforeach
@@ -67,10 +80,5 @@
         @endif
     </div>
 
-    <div class="container col-lg-6">
-        @isset($add)
-            @include('equipe.add')
-        @endisset
-    </div>
 
 @endsection
